requirements/base.txt
Django == 1.10.6
django-crispy-forms == 1.6.1
